--- a/lib/concurrent/atomic/count_down_latch.rb
+++ b/lib/concurrent/atomic/count_down_latch.rb
@@ -58,7 +58,7 @@ module Concurrent
       synchronize { @count }
     end
 
-    private
+    protected
 
     def ns_initialize(count)
       @count = count