Creates a new `ExecutorService` and related series of modules and classes through which all executors use the new synchronization. Additionally:

* Partially addresses #278, #242, and #241
* Renames `PerThreadExecutor` to `SimpleExecutorService`
* Improves the performance of Java-optimized thread pools (#280)
* Updates many classes to use `ns_initialize`
* Updates `IVar` to better support subclasses using `ns_initialize`