@jdantonio 2015-04-28
@pitr-ch This PR incorporates your suggestions. I quickly made the updates before work this morning so I expect I probably missed much.