@jdantonio 2015-04-29
@pitr-ch I consider this a spike and not ready to merge, but please provide your thoughts. I will be working on it more this evening. I prefer this approach over the other two spikes but we should explicitly choose one of the three approaches. I also want to incorporate the updates @headius suggested in #280  and incorporate #284 once it's been merged to master. I was unable to fully review #284 last night (I gave a presentation to a small group of students at the University of Akron) but I plan to review it more fully today. My initial impression is that #284 should go into master first.