@jdantonio 2015-05-02
The last update implements the performance improvements from #280. The performance difference is significant:

Before:

```
$ ./examples/benchmark_thread_pool_implementations.rb
Rehearsal ----------------------------------------------------------------------
Concurrent::JavaThreadPoolExecutor  12.200000   0.820000  13.020000 (  4.115000)
------------------------------------------------------------ total: 13.020000sec

                                         user     system      total        real
Concurrent::JavaThreadPoolExecutor   8.040000   0.320000   8.360000 (  2.734000)
```

After:
```
$ ./examples/benchmark_thread_pool_implementations.rb
Rehearsal ----------------------------------------------------------------------
Concurrent::JavaThreadPoolExecutor   2.750000   0.320000   3.070000 (  0.951000)
------------------------------------------------------------- total: 3.070000sec

                                         user     system      total        real
Concurrent::JavaThreadPoolExecutor   0.610000   0.060000   0.670000 (  0.339000)
```