@jdantonio 2015-05-03
## JRuby Executor Performance

See #280 

### Old Implementation

#### jruby-1.7.18

```
Rehearsal ------------------------------------------------------------------------
Concurrent::JavaCachedThreadPool      11.990000   1.380000  13.370000 (  8.653000)
Concurrent::JavaFixedThreadPool        8.400000   0.990000   9.390000 (  6.394000)
Concurrent::JavaSingleThreadExecutor   6.870000   0.100000   6.970000 (  4.003000)
Concurrent::JavaThreadPoolExecutor     6.590000   0.170000   6.760000 (  5.311000)
-------------------------------------------------------------- total: 36.490000sec

                                           user     system      total        real
Concurrent::JavaCachedThreadPool       7.410000   1.140000   8.550000 (  5.772000)
Concurrent::JavaFixedThreadPool        7.250000   1.020000   8.270000 (  5.242000)
Concurrent::JavaSingleThreadExecutor   4.910000   0.430000   5.340000 (  3.919000)
Concurrent::JavaThreadPoolExecutor     3.850000   0.670000   4.520000 (  3.792000)
```

#### jruby-head

```
Rehearsal ------------------------------------------------------------------------
Concurrent::JavaCachedThreadPool      10.860000   1.270000  12.130000 (  6.242069)
Concurrent::JavaFixedThreadPool        8.030000   0.880000   8.910000 (  4.917759)
Concurrent::JavaSingleThreadExecutor   6.160000   0.290000   6.450000 (  3.971956)
Concurrent::JavaThreadPoolExecutor     4.220000   0.570000   4.790000 (  4.240146)
-------------------------------------------------------------- total: 32.280000sec

                                           user     system      total        real
Concurrent::JavaCachedThreadPool       4.150000   0.840000   4.990000 (  4.100592)
Concurrent::JavaFixedThreadPool        3.860000   0.830000   4.690000 (  3.580397)
Concurrent::JavaSingleThreadExecutor   3.840000   0.630000   4.470000 (  3.355992)
Concurrent::JavaThreadPoolExecutor     3.760000   0.560000   4.320000 (  3.441682)
```

### New Implementation

#### jruby-1.7.18

```
Rehearsal ------------------------------------------------------------------------
Concurrent::JavaCachedThreadPool       2.980000   0.840000   3.820000 (  1.982000)
Concurrent::JavaFixedThreadPool        1.330000   0.450000   1.780000 (  0.919000)
Concurrent::JavaSingleThreadExecutor   0.570000   0.060000   0.630000 (  0.437000)
Concurrent::JavaThreadPoolExecutor     0.750000   0.270000   1.020000 (  0.761000)
--------------------------------------------------------------- total: 7.250000sec

                                           user     system      total        real
Concurrent::JavaCachedThreadPool       1.080000   0.680000   1.760000 (  1.038000)
Concurrent::JavaFixedThreadPool        0.920000   0.620000   1.540000 (  0.915000)
Concurrent::JavaSingleThreadExecutor   0.590000   0.340000   0.930000 (  0.543000)
Concurrent::JavaThreadPoolExecutor     0.970000   0.250000   1.220000 (  0.693000)
```

#### jruby-head

```
Rehearsal ------------------------------------------------------------------------
Concurrent::JavaCachedThreadPool       3.460000   0.880000   4.340000 (  2.240732)
Concurrent::JavaFixedThreadPool        1.040000   0.260000   1.300000 (  0.666024)
Concurrent::JavaSingleThreadExecutor   0.760000   0.080000   0.840000 (  0.516168)
Concurrent::JavaThreadPoolExecutor     0.520000   0.150000   0.670000 (  0.429587)
--------------------------------------------------------------- total: 7.150000sec

                                           user     system      total        real
Concurrent::JavaCachedThreadPool       0.750000   0.190000   0.940000 (  0.681543)
Concurrent::JavaFixedThreadPool        0.820000   0.710000   1.530000 (  0.902916)
Concurrent::JavaSingleThreadExecutor   0.390000   0.150000   0.540000 (  0.328649)
Concurrent::JavaThreadPoolExecutor     0.460000   0.300000   0.760000 (  0.602376)
```